modularize: power setters and utilities
