--- src/data/DataStructures.cpp
@@ -1,6 +1,0 @@
-#include "DataStructures.h"
-
-// Global data structures
-PowerData PhasePower[3];
-EnergyData PhaseEnergy[3];
-String serJsonResponse;
